Various cleanups to get full building across projects and platforms
coding/warnings.txt: some clarifications
projects/assets/components/dsp_comps/cic_dec.test/generate.py: python 3 update
projects/assets_ts/components/cic_dec_ts.test/generate.py: ditto
projects/assets_ts/components/downsample_complex_ts.test/generate.py: ditto
projects/assets_ts/components/fir_complex_sse_ts.test/generate.py: ditto
projects/assets/components/dsp_comps/complex_mixer_cores.hdl/Makefile: add xsim
projects/assets/hdl/primitives/zynq/Makefile: use updated macro names
runtime/dataplane/transport/impl/include/OcpiInputBuffer.h: llvm warning
runtime/hdl-support/xilinx/xilinx.mk: remove debug, allow no-license webpack
-- treat SDK and Vivao as peers like they actually are
runtime/hdl/src/HdlLSimDevice.cxx: comments
runtime/util/property/include/OcpiUtilPort.h: minor update for ultrascale zynq DMA
runtime/util/time/include/OcpiTime.h: ditto
tools/cdk/include/hdl/hdl-core.mk: indent, add missing/broken dependency
tools/cdk/include/hdl/hdl-pre.mk: rermove overly general HdlInstallDir setting
tools/cdk/python/_opencpi/util/project.py: replace non-portable mknod call (macos)
tools/cdk/scripts/setup-prerequisite.sh: do not mess with CRLFs in prereq clones

+++ projects/assets_ts/components/fir_complex_sse_ts.test/generate.py
@@ -60,7 +60,7 @@ else:
     print('  FAILED: Unsupported coefficient width: ' + str(COEFF_WIDTH))
     sys.exit(1)        
 
-num_impulse_arrays_required = min_samples_to_generate // num_taps + 1
+num_impulse_arrays_required = int(min_samples_to_generate // num_taps + 1)
 
 out_data = np.zeros(num_impulse_arrays_required*num_taps, dtype=utu.dt_iq_pair)
 out_data['real_idx'] = np.tile(impulse_array, num_impulse_arrays_required)
@@ -68,7 +68,7 @@ out_data['imag_idx'] = out_data['real_idx']
 
 # Write to file
 message_size = 2048 #This is the maximum allowed by current buffer negotiation system
-samples_per_message = message_size / bytes_per_sample
+samples_per_message = int(message_size / bytes_per_sample)
 with open(sys.argv[1], 'wb') as f:
     utu.add_msg(f, iqm.INTERVAL_OPCODE, array.array('I',(int('00000000',16), int('00001FFF',16)))) #8191
     utu.add_msg(f, iqm.TIME_OPCODE, array.array('I',(int('0000AAAA',16), int('0000BBBB',16))))
